Changed customOrderId to long everywhere due to using date for receipt's customOrderId. Removed onSwap listeners. Dragging responsibility is divided to base DraggableListFragment and custom DraggableEditableListFragment. Added receipt's ordering support to OrderingPreferencesManager. Now every new receipt if ordering is turned on gets customOrderId = receipt.getDate() + currentTime, because date has no time information.

+++ app/src/main/java/co/smartreceipts/android/model/ColumnDefinitions.java
@@ -15,12 +15,13 @@ public interface ColumnDefinitions<T> {
      * Gets a {@link Column} instance for a specific definition (i.e. column
      * name). Please note that different definition names can potentially be applied to the same column).
      *
-     * @param id the unique identifier for the column
+     * @param id             the unique identifier for the column
      * @param definitionName the name of the new column
-     * @param syncState the current {@link SyncState} of the column
+     * @param syncState      the current {@link SyncState} of the column
+     * @param customOrderId  the order id for this column
      * @return a new column instance or {@code null} if none can be found
      */
-    Column<T> getColumn(int id, @NonNull String definitionName, @NonNull SyncState syncState, int customOrderId);
+    Column<T> getColumn(int id, @NonNull String definitionName, @NonNull SyncState syncState, long customOrderId);
 
     /**
      * Gets a list of all {@link Column} instances that are available as part
