+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/SettingUserIdColumn.java
@@ -13,7 +13,8 @@ public final class SettingUserIdColumn<T> extends AbstractColumnImpl<T> {
 
     private final UserPreferenceManager mPreferences;
 
-    public SettingUserIdColumn(int id, @NonNull String name, @NonNull SyncState syncState, @NonNull UserPreferenceManager preferences, int customOrderId) {
+    public SettingUserIdColumn(int id, @NonNull String name, @NonNull SyncState syncState,
+                               @NonNull UserPreferenceManager preferences, long customOrderId) {
         super(id, name, syncState, customOrderId);
         mPreferences = preferences;
     }
