+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptCategoryCodeColumn.java
@@ -11,7 +11,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReceiptCategoryCodeColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReceiptCategoryCodeColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptCategoryCodeColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
