+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptCommentColumn.java
@@ -11,7 +11,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReceiptCommentColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReceiptCommentColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptCommentColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
