+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptExtra3Column.java
@@ -11,7 +11,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReceiptExtra3Column extends AbstractColumnImpl<Receipt> {
 
-    public ReceiptExtra3Column(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptExtra3Column(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
