+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptIdColumn.java
@@ -11,7 +11,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReceiptIdColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReceiptIdColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptIdColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
