+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptNameColumn.java
@@ -15,7 +15,7 @@ public final class ReceiptNameColumn extends AbstractColumnImpl<Receipt> {
         super(id, name, syncState);
     }
 
-    public ReceiptNameColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptNameColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
