+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptNetExchangedPricePlusTaxColumn.java
@@ -20,7 +20,7 @@ public final class ReceiptNetExchangedPricePlusTaxColumn extends AbstractExchang
     private final UserPreferenceManager mPreferences;
 
     public ReceiptNetExchangedPricePlusTaxColumn(int id, @NonNull String name, @NonNull SyncState syncState,
-                                                 @NonNull Context context, @NonNull UserPreferenceManager preferences, int customOrderId) {
+                                                 @NonNull Context context, @NonNull UserPreferenceManager preferences, long customOrderId) {
         super(id, name, syncState, context, customOrderId);
         mPreferences = preferences;
     }
