+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReceiptPaymentMethodColumn.java
@@ -11,7 +11,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReceiptPaymentMethodColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReceiptPaymentMethodColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReceiptPaymentMethodColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
