+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReportCommentColumn.java
@@ -13,7 +13,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReportCommentColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReportCommentColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReportCommentColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
