+++ app/src/main/java/co/smartreceipts/android/model/impl/columns/receipts/ReportCostCenterColumn.java
@@ -13,7 +13,7 @@ import co.smartreceipts.android.sync.model.SyncState;
  */
 public final class ReportCostCenterColumn extends AbstractColumnImpl<Receipt> {
 
-    public ReportCostCenterColumn(int id, @NonNull String name, @NonNull SyncState syncState, int customOrderId) {
+    public ReportCostCenterColumn(int id, @NonNull String name, @NonNull SyncState syncState, long customOrderId) {
         super(id, name, syncState, customOrderId);
     }
