+++ app/src/main/java/co/smartreceipts/android/persistence/database/tables/adapters/ReceiptDatabaseAdapter.java
@@ -106,7 +106,7 @@ public final class ReceiptDatabaseAdapter implements SelectionBackedDatabaseAdap
         final String extra_edittext_1 = cursor.getString(extra_edittext_1_Index);
         final String extra_edittext_2 = cursor.getString(extra_edittext_2_Index);
         final String extra_edittext_3 = cursor.getString(extra_edittext_3_Index);
-        final int orderId = cursor.getInt(orderIdIndex);
+        final long orderId = cursor.getLong(orderIdIndex);
         File file = null;
         if (!TextUtils.isEmpty(path) && !DatabaseHelper.NO_DATA.equals(path)) {
             file = mStorageManager.getFile(trip.getDirectory(), path);
